Add rule

+++ src/rules/git_rm_recursive.rs
@@ -1,37 +1,39 @@
 use crate::{
     cli::command::CrabCommand,
-    rules::{match_rule_with_git_support, utils::git::get_new_command_with_git_support},
-    Rule,
+    rules::{
+        utils::git::{get_new_command_with_git_support, match_rule_with_git_support},
+        Rule,
+    },
+    shell::Shell,
 };
-use shell::Shell;
 
 fn auxiliary_match_rule(command: &CrabCommand) -> bool {
-    command.script.contains(" rm ")
-        && command.output.as_ref().map_or(false, |o| {
-            o.contains("fatal: not removing '") && o.contains("' recursively without -r")
-        })
+    if let Some(stdout) = &command.output {
+        command.script.contains(" rm ")
+            && stdout.contains("fatal: not removing '")
+            && stdout.contains("' recursively without -r")
+    } else {
+        false
+    }
 }
 
-pub fn match_rule(command: &mut CrabCommand, _system_shell: Option<&dyn Shell>) -> bool {
+pub fn match_rule(command: &mut CrabCommand, system_shell: Option<&dyn Shell>) -> bool {
     match_rule_with_git_support(auxiliary_match_rule, command)
 }
 
 fn auxiliary_get_new_command(
     command: &CrabCommand,
-    _system_shell: Option<&dyn Shell>,
+    system_shell: Option<&dyn Shell>,
 ) -> Vec<String> {
-    let mut command_parts = command.script.split_whitespace().collect::<Vec<&str>>();
-    if let Some(index) = command_parts.iter().position(|&r| r == "rm") {
-        command_parts.insert(index + 1, "-r");
+    let mut command_parts = command.script_parts.clone();
+    if let Some(index) = command_parts.iter().position(|r| r == "rm") {
+        command_parts.insert(index + 1, "-r".to_string());
     }
     vec![command_parts.join(" ")]
 }
 
-pub fn get_new_command(
-    command: &mut CrabCommand,
-    _system_shell: Option<&dyn Shell>,
-) -> Vec<String> {
-    get_new_command_with_git_support(auxiliary_get_new_command, command, _system_shell)
+pub fn get_new_command(command: &mut CrabCommand, system_shell: Option<&dyn Shell>) -> Vec<String> {
+    get_new_command_with_git_support(auxiliary_get_new_command, command, system_shell)
 }
 
 pub fn get_rule() -> Rule {
@@ -54,7 +56,7 @@ mod tests {
     use rstest::rstest;
 
     #[rstest]
-    #[case("git rm foo", "fatal: not removing 'foo' recursively without -r", true)]
+    #[case("git rm fo", "fatal: not removing 'foo' recursively without -r", true)]
     #[case(
         "git rm foo bar",
         "fatal: not removing 'foo bar' recursively without -r",
